fix: Gen 2 — soundness, network hardening, async fixes, lint deep clean, dead code purge, build optimization
Track A — Soundness:
- Add SAFETY comments to unsafe impl Send/Sync on dynosaur types
- Fix TOCTOU key file permissions (atomic mode at creation)
- Fix transmute endianness in JobId (explicit le_bytes conversion)
- Serialize tests with env var mutation to prevent data races

Track B — Network Hardening:
- Add per-peer rate limiting (100msg/10s sliding window) to P2P protocol
- Bound bincode deserialization to MAX_MESSAGE_SIZE (16MB cap)
- Replace SSH command denylist with allowlist of permitted deployment ops

Track C — Async & Concurrency:
- Replace QoS 30s busy-wait polling with tokio::sync::Notify
- Add backpressure cap (1024 concurrent jobs) to producer pipeline

Track D — Clippy Deep Clean:
- Remove 23 unjustified clippy allows (100 → 77 workspace-level)
- Fix 18 inline clippy warnings across 13 files
- Now enforcing: borrowed_box, float_cmp, inefficient_to_string, len_zero, useless_conversion

Track E — TODO/Dead Code Purge:
- Resolve 7 TODOs (including hardcoded service_id/blueprint_id in QoS)
- Delete 10 stale TODO/FIXME comments
- Remove 5 dead functions from macros, std, manager
- 24 TODOs kept (genuine missing features with context)

Track F — Build Optimization:
- Populate workspace-hack via cargo-hakari (22 → 534 lines)
- Add workspace-hack dep to all 95 workspace members
- Audit: 2 meta-crates identified as INLINE candidates (metrics, stores)
- Audit: ~70% of non-default feature flags have zero CI coverage

+++ crates/clients/tangle/src/client.rs
@@ -381,6 +381,7 @@ pub struct TangleClient {
     block_subscription: Arc<Mutex<Option<u64>>>,
 }
 
+#[allow(clippy::missing_fields_in_debug)] // provider/signer/subscription intentionally omitted
 impl fmt::Debug for TangleClient {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.debug_struct("TangleClient")
@@ -968,7 +969,7 @@ impl TangleClient {
             restaking: RestakingMetadata {
                 stake: restaking_meta.stake,
                 delegation_count: restaking_meta.delegationCount,
-                status: RestakingStatus::from(u8::from(restaking_meta.status)),
+                status: RestakingStatus::from(restaking_meta.status),
                 leaving_round: restaking_meta.leavingRound,
             },
         })
@@ -1455,7 +1456,7 @@ impl TangleClient {
         Ok(RestakingMetadata {
             stake: restaking_meta.stake,
             delegation_count: restaking_meta.delegationCount,
-            status: RestakingStatus::from(u8::from(restaking_meta.status)),
+            status: RestakingStatus::from(restaking_meta.status),
             leaving_round: restaking_meta.leavingRound,
         })
     }
@@ -1538,7 +1539,7 @@ impl TangleClient {
             .call()
             .await
             .map_err(|e| Error::Contract(e.to_string()))?;
-        Ok(DelegationMode::from(u8::from(mode)))
+        Ok(DelegationMode::from(mode))
     }
 
     /// Check if delegator is whitelisted for operator.
